third_party/blink/web_tests/external/wpt/css/css-contain/content-visibility/content-visibility-070-expected.txt

This is a testharness.js-based test.
FAIL Content Visibility: off-screen selection assert_equals: step6 height expected 10 but got 100
Harness: the test ran to completion.
